Factor out requirement components in RequirementFulfillment (4/n)
